gamescore invitecode

+++ src/entity/gamescore.rs
@@ -8,10 +8,10 @@ pub struct Model {
     #[sea_orm(primary_key)]
     pub id: i32,
     pub created_at: i32,
-    pub updated_at: i32,
+    pub updated_at: i32,    
     pub user_id: i64,
     pub game_type: String,
-    pub game_level: i32,
+    pub game_level: i8,
     pub score: i32,
 }
